requirements.txt
h5py==3.2.1
hydra-core==1.1.1
numpy==1.22.3
torch==1.11.0
pytorch-lightning==1.4.2
torchmetrics==0.6.0
tqdm==4.62.3
matplotlib==3.4.3
scikit-learn==0.24.2
scipy==1.7.1

#changed by Rohit
setuptools==65.5.0
wheel==0.38.4
gym==0.21.0